requirements.txt
google-cloud-texttospeech==2.27.0
pydub==0.25.1